Refactor library into modules and fix deprecations
